+++ react-ui/src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import { HomePage } from "./components/Home-page/Home-page";
-import { RestaurantInfo } from './components/Restaurant-info/Restaurant-info';
-import { NewRestaurant } from './components/New-Restaurant/New-Restaurant';
-import { Location } from './components/Location/Location';
+import { HomePage } from "./components/Homepage/Homepage";
+import { RestaurantInfo } from './components/Info/RestaurantInfo';
+import { NewRestaurant } from './components/New/NewRestaurant';
+import { Search } from './components/Search/Search';
 import {
   BrowserRouter as Router,
   Switch,
@@ -15,7 +15,7 @@ function App() {
       <Switch>
         <Route exact path="/" component={HomePage} />
         <Route exact path="/restaurants/:id" component={RestaurantInfo}/>
-        <Route exact path="/restaurants/location/:location" component={Location}/>
+        <Route exact path="/restaurants/location/:location" component={Search}/>
         <Route path="/add" component={NewRestaurant}/>
       </Switch>
     </Router>
