+++ react-ui/src/components/Homepage/Homepage.css
@@ -21,7 +21,6 @@
     font-weight: 400; 
     line-height: 26.4px; 
     margin-bottom: 30px;
-    text-decoration: none;
   } 
   
   h3 {
