+++ react-ui/src/components/Info/RestaurantInfo.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import Button from '@material-ui/core/Button';
 import { Link } from 'react-router-dom'
-import './Restaurant-info.css';
+import './RestaurantInfo.css';
 
 export const RestaurantInfo = (props) => {
     const {id} = props.match.params
