+++ react-ui/src/components/Search/LocationSearch.js
